% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_columns}
\alias{assert_columns}
\title{Assert that args in ellipses are columns in \code{df}}
\usage{
assert_columns(df, ...)
}
\arguments{
\item{df}{A data frame.}

\item{...}{Column names to test.}
}
\description{
Assert that args in ellipses are columns in \code{df}
}
