% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_columns_unique}
\alias{assert_columns_unique}
\title{Assert that column names are not identical}
\usage{
assert_columns_unique(...)
}
\arguments{
\item{...}{Column names}
}
\description{
Takes in any number of column name variables, asserts they are not the same.
}
