% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_error_correct_avg_trend}
\alias{assert_error_correct_avg_trend}
\title{If using \code{error_correct}, then check that the columns are either in the
\code{formula_vars} or \code{average_cols}, otherwise produce an error.}
\usage{
assert_error_correct_avg_trend(
  formula_vars,
  average_cols,
  error_correct,
  error_correct_cols
)
}
\arguments{
\item{formula_vars}{Variables included in the model formula, generated by
\code{all.vars(formula)}.}

\item{average_cols}{Column name(s) of column(s) for use in grouping data for averaging,
such as regions. If missing, uses global average of the data for infilling.}

\item{error_correct}{Logical value indicating whether or not whether mean error
should be used to adjust predicted values. If \code{TRUE}, the mean error between
observed and predicted data points will be used to adjust predictions. If
\code{error_correct_cols} is not \code{NULL}, mean error will be used within those
groups instead of overall mean error.}

\item{error_correct_cols}{Column names of data frame to group by when applying
error correction to the predicted values.}
}
\description{
If using \code{error_correct}, then check that the columns are either in the
\code{formula_vars} or \code{average_cols}, otherwise produce an error.
}
