% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_function}
\alias{assert_function}
\title{Assert that \code{x} is a function}
\usage{
assert_function(x)
}
\arguments{
\item{x}{Supposed function to test}
}
\description{
Assert that \code{x} is a function
}
