% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_group_sort_col}
\alias{assert_group_sort_col}
\title{Assert sort column for use in average trend functions}
\usage{
assert_group_sort_col(formula_vars, average_cols, sort_col)
}
\arguments{
\item{formula_vars}{Variables included in the model formula, generated by
\code{all.vars(formula)}.}

\item{average_cols}{Column name(s) of column(s) for use in grouping data for averaging,
such as regions. If missing, uses global average of the data for infilling.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series.}
}
\value{
Column name for sorting averaged data to generate trend.
}
\description{
Takes in formula variables and average columns, as well as sort column argument.
If \code{sort_col} is provided, an error is generated if it is not
in \code{formula_vars} or \code{average_cols}. If \code{sort_col} is not provided, then
warnings are generated that no average trend can be generated.
}
