% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_h}
\alias{assert_h}
\title{Assert that h, for forecasting, is > 0}
\usage{
assert_h(h)
}
\arguments{
\item{h}{Number of missing values in series passed to forecast}
}
\description{
Assert that h, for forecasting, is > 0
}
