% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_inla}
\alias{assert_inla}
\title{Assert if INLA is installed, for use in \code{predict_inla...()} functions.}
\usage{
assert_inla()
}
\description{
Assert if INLA is installed, for use in \code{predict_inla...()} functions.
}
