% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_model}
\alias{assert_model}
\title{Assert that \code{x} is a function}
\usage{
assert_model(x)
}
\arguments{
\item{x}{Model function to test}
}
\description{
Assert that \code{x} is a function
}
