% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_numeric}
\alias{assert_numeric}
\title{Assert numeric value}
\usage{
assert_numeric(x, n)
}
\arguments{
\item{x}{Numeric value to check}

\item{n}{Required length of vector}
}
\description{
Assert numeric value
}
