% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_numeric_cols}
\alias{assert_numeric_cols}
\title{Assert columns in \code{df} are numeric}
\usage{
assert_numeric_cols(cols, df)
}
\arguments{
\item{cols}{Columns to check}

\item{df}{Data frame with columns}
}
\description{
Assert columns in \code{df} are numeric
}
