% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_numeric_cols_avg}
\alias{assert_numeric_cols_avg}
\title{Assert columns in \code{df} are numeric, for use with average trend functions}
\usage{
assert_numeric_cols_avg(cols, df)
}
\arguments{
\item{cols}{Columns to check}

\item{df}{Data frame with columns}
}
\description{
Checks that columns are numeric in \code{df}, and returns error message specifying that
they must either be numeric or explicitly included in the \code{average_cols} grouping
to be used for averaging.
}
