% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_asserts.R
\name{assert_test_col}
\alias{assert_test_col}
\title{Assert that test_col is of logical type}
\usage{
assert_test_col(df, test_col)
}
\arguments{
\item{df}{Data frame of model data.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored.}
}
\description{
Assert that test_col is of logical type
}
