% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augury_add_columns.R
\name{augury_add_columns}
\alias{augury_add_columns}
\title{Adds empty columns to df}
\usage{
augury_add_columns(df, cols)
}
\arguments{
\item{df}{Data frame}

\item{cols}{Vector of column names to add to df}
}
\description{
Used when models are not generated based on \code{obs_filter}.
}
