% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_aarr.R
\name{calculate_aarr}
\alias{calculate_aarr}
\title{Extract AARR from vector of years and prevalence}
\usage{
calculate_aarr(years, prevalence)
}
\arguments{
\item{years}{Vector of year values}

\item{prevalence}{Vector of prevalence values}
}
\description{
Extract AARR from vector of years and prevalence
}
