% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_error.R
\name{calculate_sq_ch}
\alias{calculate_sq_ch}
\title{Calculate change error}
\usage{
calculate_sq_ch(response, pred, test_period, test_period_flex)
}
\arguments{
\item{response}{Observations to test.}

\item{pred}{Predictions to test.}

\item{test_period}{Length of period to test for RMChE. If \code{NULL}, beginning and end
points of each group in \code{group_col} are compared. Otherwise, \code{test_period} must
be set to an integer \code{n} and for each group, comparisons are made between
the end point and \code{n} periods prior.}

\item{test_period_flex}{Logical value indicating if \code{test_period} is less than
the full length of the series, should change error still be calculated for that
point. Defaults to \code{FALSE}.}
}
\description{
For use to calculate RMChE, this calculates the squared change error between observed and predicted columns.
}
