% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_correct.R
\name{error_correct_fn}
\alias{error_correct_fn}
\title{Use mean error to correct predictions}
\usage{
error_correct_fn(
  df,
  response,
  pred_col,
  upper_col,
  lower_col,
  test_col,
  error_correct,
  error_correct_cols
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{response}{Column name of response variable.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored. See \code{\link[=model_error]{model_error()}}
for details on the methods and metrics returned.}

\item{error_correct}{Logical value indicating whether or not whether mean error
should be used to adjust predicted values. If \code{TRUE}, the mean error between
observed and predicted data points will be used to adjust predictions. If
\code{error_correct_cols} is not \code{NULL}, mean error will be used within those
groups instead of overall mean error.}

\item{error_correct_cols}{Column names of data frame to group by when applying
error correction to the predicted values.}
}
\value{
Depending on the value passed to \code{ret}, either a data frame with
predicted data, a vector of errors from \code{\link[=model_error]{model_error()}}, a fitted model, or a list with all 3.
}
\description{
Use mean error to correct predictions
}
