% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_df.R
\name{expand_df}
\alias{expand_df}
\title{Expand input data to make explicit missing values}
\usage{
expand_df(
  df,
  ...,
  response = "value",
  keep_no_obs = TRUE,
  keep_before_obs = FALSE,
  sort_col = "year",
  sort_descending = FALSE,
  group_col = "iso3",
  join_covariates = FALSE
)
}
\arguments{
\item{df}{Data frame.}

\item{...}{Named vectors to pass to expand grid.}

\item{response}{Column name of response variables whose missing values will be
infilled and projected, defaults to \code{"value"}.}

\item{keep_no_obs}{Logical value indicating whether or not to keep rows in the
expanded data frame when there is no data. Defaults to \code{TRUE}. This is done based
on the \code{group_col}, if provided.}

\item{keep_before_obs}{Logical value indicating when data is available, whether or not
to keep rows in the expanded data frame that lie before the first observed point.
Defaults to \code{FALSE}. This is done based on the \code{sort_col} and \code{group_col}, if provided.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series. If \code{NULL}, not used. Defaults to \code{"year"}.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.
If \code{NULL}, not used. Defaults to \code{"iso3"}.}

\item{join_covariates}{Logical value indicating whether or not to join the final
expanded data frame to the \link{covariates_df} data frame. If \code{TRUE},
\code{iso3} and \code{year} must be columns within the input \code{df}.}
}
\value{
Expanded data frame with explicit missing values.
}
\description{
\code{expand_df()} is a wrapper around \code{\link[tidyr:expand_grid]{tidyr::expand_grid()}} and \code{\link[dplyr:mutate-joins]{dplyr::right_join()}}
that can be used to make missing values explicit within a data frame prior to
it being passed to a \code{predict_...()} function.
}
