% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_df.R
\name{expand_df_min}
\alias{expand_df_min}
\title{Helper for \code{expand_df_filter()} to calculate min for keeping data}
\usage{
expand_df_min(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Value.
}
\description{
Want to make sure that this returns \code{-Inf} rather than \code{Inf} for the \code{expand_df_filter()}
filter to work
}
