% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_data.R
\name{filter_model_data}
\alias{filter_model_data}
\title{Filters data for modeling}
\usage{
filter_model_data(df, formula_vars, filter_na)
}
\arguments{
\item{df}{Data frame of model data.}

\item{formula_vars}{Character vector of variables used in the model. Can be
extracted from a formula using \code{all.vars(fmla)}.}

\item{filter_na}{Character value specifying how, if at all, to filter \code{NA}
values from the dataset prior to applying the model.  By default, all
observations with missing values are removed, although it can also remove
rows only if they have missing dependent or independent variables, or no
filtering at all.}
}
\value{
A data frame.
}
\description{
Based on the \code{filter_na} argument, data is filtered for modelling based on
\code{NA} values within \code{df}.
}
