% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_aarr.R
\name{fit_aarr_model}
\alias{fit_aarr_model}
\title{Generate prediction from model object}
\usage{
fit_aarr_model(
  df,
  response,
  interpolate,
  test_col,
  sort_col,
  sort_descending,
  sort_col_min,
  group_col,
  group_models,
  obs_filter,
  pred_col
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{response}{Column name of prevalence variable to be used to calculate
AARR.}

\item{interpolate}{Logical value, whether or not to interpolate values based on
estimated AARR between observations. Defaults to \code{FALSE}.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored. See \code{\link[=model_error]{model_error()}}
for details on the methods and metrics returned.}

\item{sort_col}{Column name of column to arrange data by in \code{dplyr::arrange()},
prior to filtering for latest contiguous time series and producing the
forecast. Not used if \code{NULL}, defaults to \code{"year"}.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}

\item{sort_col_min}{If provided, a numeric value that sets a minimum value needed
to be met in the \code{sort_col} for an observation to be used in calculating AARR.
If \code{sort_col = "year"} and \code{sort_col_min = 2008}, then only observations
from 2008 onward will be used in calculating AARR.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.
If \code{NULL}, not used. Defaults to \code{"iso3"}.}

\item{group_models}{Logical, if \code{TRUE}, fits and predicts models individually onto
each \code{group_col}. If \code{FALSE}, a general model is fit across the entire data
frame.}

\item{obs_filter}{String value of the form "\verb{logical operator} \code{integer}"
that specifies the number of observations required to fit the model and
replace observations with predicted values. This is done in
conjunction with \code{group_col}. So, if \code{group_col = "iso3"} and
\code{obs_filter = ">= 5"}, then for this model, predictions will only be used
for \code{iso3} vales that have 5 or more observations. Possible logical operators
to use are \code{>}, \code{>=}, \code{<}, \code{<=}, \code{==}, and \code{!=}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{If `group_models = FALSE`, then `obs_filter` is only used to determine when
predicted values replace observed values but **is not** used to restrict values
from being used in model fitting. If `group_models = TRUE`, then a model
is only fit for a group if they meet the `obs_filter` requirements. This provides
speed benefits, particularly when running INLA time series using `predict_inla()`.
}\if{html}{\out{</div>}}}

\item{pred_col}{Column name to store predicted value.}
}
\value{
A data frame.
}
\description{
\code{fit_aarr_data()} calculates AARR and then generates a prediction based on calculated AARR.
}
