% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_general_mdl_avg_trend.R
\name{fit_general_average_model}
\alias{fit_general_average_model}
\title{Fit general model to averages and apply trend to original data}
\usage{
fit_general_average_model(
  df,
  model,
  formula,
  average_cols,
  weight_col,
  ...,
  formula_vars,
  test_col,
  group_col,
  group_models,
  sort_col,
  pred_col,
  upper_col,
  lower_col,
  filter_na,
  ret,
  error_correct,
  error_correct_cols,
  shift_trend
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{model}{An R function that outputs a model object with a \code{predict.model} generic,
where \code{\link[stats:family]{stats::family()}} contains an inverse link function \code{linkinv} and
\code{predict.model()} accepts the \code{se.fit = TRUE} argument and returns confidence
intervals. This includes \link[stats:lm]{stats::lm}, \link[stats:glm]{stats::glm}, and \link[lme4:lmer]{lme4::lmer}.}

\item{formula}{A formula that will be supplied to the model, such as \code{y~x}.
Variables defined in the formula will be used in the averaging. If the
variable is defined as part of \code{average_cols}, then it will be used within
\code{\link[dplyr:group_by]{dplyr::group_by()}} prior to averaging. If it is not a part of \code{average_cols},
then it must be a numeric column whose average will be taken.}

\item{average_cols}{Column name(s) of column(s) for use in grouping data for averaging,
such as regions. If missing, uses global average of the data for infilling.}

\item{weight_col}{Column name of column of weights to be used in averaging, such
as country population.}

\item{...}{Additional arguments passed to \code{\link[INLA:inla]{INLA::inla()}}.}

\item{formula_vars}{Variables included in the model formula, generated by
\code{all.vars(formula)}.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored. See \code{\link[=model_error]{model_error()}}
for details on the methods and metrics returned.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.}

\item{group_models}{Logical, whether or not to run separate models for each group
defined by \code{average_cols}. If the \code{sort_col} is part of \code{average_cols}, then
it is not used to group models.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}

\item{filter_na}{Character value specifying how, if at all, to filter \code{NA}
values from the dataset prior to applying the model. By default, only
observations with missing predictors are removed, although it can also remove
rows only if they have missing dependent or independent variables, or no
filtering at all. Model prediction and fitting are done in one pass with
\code{\link[INLA:inla]{INLA::inla()}}, so there will be no predictions if observations with
missing dependent variables are removed.}

\item{ret}{Character vector specifying what values the function returns. Defaults
to returning a data frame, but can return a vector of model error, the
model itself or a list with all 3 as components.}

\item{error_correct}{Logical value indicating whether or not whether mean error
should be used to adjust predicted values. If \code{TRUE}, the mean error between
observed and predicted data points will be used to adjust predictions. If
\code{error_correct_cols} is not \code{NULL}, mean error will be used within those
groups instead of overall mean error.}

\item{error_correct_cols}{Column names of data frame to group by when applying
error correction to the predicted values.}

\item{shift_trend}{Logical value specifying whether or not to shift predictions
so that the trend matches up to the last observation. If \code{error_correct} and
\code{shift_trend} are both \code{TRUE}, \code{shift_trend} takes precedence.}
}
\value{
List of \code{mdl} (fitted model) and \code{df} (data frame with fitted values
and confidence bounds generated from the model).
}
\description{
Used within \code{predict_general_mdl_avg_trend()}, this function fits the model to the data
frame, working whether the model is being fit across the entire data frame or
being fit to each group individually. Data is filtered prior to fitting,
model(s) are fit, and then fitted values are generated on the original.
}
\details{
If fitting models individually to each group, \code{mdl} will never be returned, as
as these are instead a large group of models. Otherwise, a list of \code{mdl} and \code{df}
is returned and used within \code{predict_general_mdl()}.
}
