% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_forecast.R
\name{forecast_series}
\alias{forecast_series}
\title{Forecast data series}
\usage{
forecast_series(x, forecast_function, ...)
}
\arguments{
\item{x}{Series to forecast, coming from \code{trim_series()}}

\item{forecast_function}{An R function that outputs a forecast object coming from the
forecast package. You can directly pass \code{\link[forecast:reexports]{forecast::forecast()}} to the
function, or you can pass other wrappers to it such as \code{\link[forecast:ses]{forecast::holt()}} or
\code{\link[forecast:ses]{forecast::ses()}}.}

\item{...}{Additional arguments passed to the forecast function.}
}
\value{
Forecast model.
}
\description{
Using series coming from \code{trim_series()}, it uses latest observed values
to forecast missing values.
}
