% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_average_trend.R
\name{get_average_df}
\alias{get_average_df}
\title{Produces averaged data frame that can then be passed for modelling.}
\usage{
get_average_df(df, cols, average_cols, weight_col = NULL)
}
\arguments{
\item{df}{Data frame of model data.}

\item{cols}{Name of numeric column(s) to take average of.}

\item{average_cols}{Column name(s) of column(s) for use in grouping data for averaging,
such as regions. If missing, uses global average of the data for infilling.}

\item{weight_col}{Column name of column of weights to be used in averaging, such
as country population.}
}
\value{
Data frame of \code{cols} averaged across \code{average_cols}
}
\description{
Produces averaged data frame that can then be passed for modelling.
}
