% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_models.R
\name{get_forecast_data}
\alias{get_forecast_data}
\title{Get data for forecast models}
\usage{
get_forecast_data(df, response, sort_col, sort_descending, test_col)
}
\arguments{
\item{df}{Data frame of model data.}

\item{response}{Column name of response variable to be used as the input to the
forecast function.}

\item{sort_col}{Column name of column to arrange data by in \code{dplyr::arrange()},
prior to filtering for latest contiguous time series and producing the
forecast.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored.}
}
\value{
A data series.
}
\description{
Keep only the latest contiguous time series, dropping all other NA values
from the response variable. Removes test column variables first.
}
