% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_average_trend.R
\name{get_formula_avg_cols}
\alias{get_formula_avg_cols}
\title{Get variables that need to be averaged from formula.}
\usage{
get_formula_avg_cols(df, formula_vars, average_cols)
}
\arguments{
\item{df}{Data frame of model data.}

\item{formula_vars}{Variables included in the model formula, generated by
\code{all.vars(formula)}.}

\item{average_cols}{Column name(s) of column(s) for use in grouping data for averaging,
such as regions. If missing, uses global average of the data for infilling.}
}
\description{
Pulls out variables not being used for grouping that will be averaged, checks
that they are numeric, and returns them.
}
