% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_wrappers.R
\name{grouped_inla_ts}
\alias{grouped_inla_ts}
\title{Use INLA for autoregressive time series prediction by group}
\usage{
grouped_inla_ts(
  df,
  group_col = "iso3",
  formula = value ~ INLA::f(year_n, model = "ar1"),
  family = "gaussian",
  control.predictor = list(compute = TRUE),
  control.inla = list(strategy = "laplace"),
  ...,
  ret = c("df", "all", "error", "model"),
  test_col = NULL,
  pred_col = "pred",
  upper_col = "upper",
  lower_col = "lower",
  filter_na = c("predictors", "response", "all", "none"),
  type_col = NULL,
  types = c("imputed", "imputed", "projected"),
  type_group = "iso3",
  type_sort = "year",
  source_col = NULL,
  source = NULL,
  replace_obs = c("missing", "all", "none"),
  error_correct = FALSE,
  error_correct_cols = NULL
)
}
\description{
Use INLA for autoregressive time series prediction by group
}
