% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_models.R
\name{inla_model}
\alias{inla_model}
\title{Bayesian analysis of structured additive models}
\usage{
inla_model(
  formula,
  data,
  family = "gaussian",
  control.predictor = list(compute = TRUE),
  control.inla = list(strategy = "laplace"),
  ...
)
}
\arguments{
\item{formula}{A \code{inla} formula like
\code{y ~ 1 + z + f(ind, model = "iid") + f(ind2, weights, model = "ar1")}.
This is much like the formula for a \code{glm} except that smooth or spatial
terms can be added to the right hand side of the formula. See \link[INLA:f]{INLA::f}
for full details.}

\item{data}{A data frame containing the variables in the model.}

\item{family}{A string indicating the likelihood family. The default is \code{gaussian} with
identity link. All valid families available through \code{model_families()}. Family
documentation available using \code{\link[INLA:doc]{INLA::inla.doc()}}.}

\item{control.predictor}{See \code{\link[INLA:control.predictor]{INLA::control.predictor()}}}

\item{control.inla}{See \code{\link[INLA:control.inla]{INLA::control.inla()}}}

\item{...}{Additional arguments passed to \code{\link[INLA:inla]{INLA::inla()}}.}
}
\value{
\code{inla_model} returns an object of class \code{"inla"}, a list containing
arguments defined at \code{\link[INLA:inla]{INLA::inla()}}.
}
\description{
Provides a convenient wrapper around the inla function from the INLA package,
used to perform Bayesian analysis of additive models using Integrated Nested
Laplace approximation. Defaults set for what is used in Billions modeling,
while allowing all arguments to be adjusted for exploration and testing.
}
\details{
See \code{\link[INLA:inla]{INLA::inla()}} for full documentation and guidance.
}
