% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_aarr.R
\name{interpolate_aarr}
\alias{interpolate_aarr}
\title{Interpolate using AARR from vector of years and prevalence}
\usage{
interpolate_aarr(years, prevalence)
}
\arguments{
\item{years}{Vector of year values}

\item{prevalence}{Vector of prevalence values}
}
\description{
Interpolate using AARR from vector of years and prevalence
}
