% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_df.R
\name{join_covariates_df}
\alias{join_covariates_df}
\title{Join data frame with covariates data frame}
\usage{
join_covariates_df(df, join_covariates)
}
\arguments{
\item{df}{Data frame.}

\item{join_covariates}{Logical value indicating whether or not to join the final
expanded data frame to the \link{covariates_df} data frame. If \code{TRUE},
\code{iso3} and \code{year} must be columns within the input \code{df}.}
}
\value{
Joined data frame.
}
\description{
Left joins a data frame with \code{augury::covariates_df}.
}
