% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_models.R
\name{likelihood_families}
\alias{likelihood_families}
\title{Retrieve all valid inla likelihood families}
\usage{
likelihood_families()
}
\value{
A character vector.
}
\description{
\code{likelihood_families()} gives a character vector of all valid likelihood families
available in \code{\link[INLA:inla]{INLA::inla()}}.
}
