% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_interpolation.R
\name{linear_interpolation}
\alias{linear_interpolation}
\title{Linearly interpolate columns}
\usage{
linear_interpolation(
  df,
  cols,
  groups = NULL,
  sort_by = NULL,
  sort_descending = FALSE
)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{Names of columns to interpolate.}

\item{groups}{Names of column to group by when interpolating, such as country ISO3 codes.}

\item{sort_by}{Names of columns to arrange data frame by prior to interpolation, such as year.}

\item{sort_descending}{Logical value on whether the sorted values should be descending. Defaults to \code{FALSE}.}
}
\value{
A data frame.
}
\description{
\code{linear_interpolation()} does simple linear interpolation on specified columns
in a data frame.
}
