% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_error.R
\name{model_error}
\alias{model_error}
\title{Get modeling error from a data frame}
\usage{
model_error(df, response, pred_col, test_col)
}
\value{
A named vector of errors: RMSE, MSE, and MASE.
}
\description{
\code{model_error()} calculates modeling error using observed and fitted values from
the data frame. If \code{test_col} is provided, the error is only calculated on
observations that were excluded from modeling for test purpose. Otherwise,
the error is calculated for all non-missing values.
}
