% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formulas.R
\name{parse_formula}
\alias{parse_formula}
\title{Asserts formula and extract variables}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{A formula that will be supplied to the model, such as \code{y~x}.}
}
\description{
Asserts formula and extract variables
}
