% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_filter.R
\name{parse_obs_filter}
\alias{parse_obs_filter}
\title{Parse obs filter intro string to be evaluated}
\usage{
parse_obs_filter(obs_filter, response)
}
\arguments{
\item{obs_filter}{Obs filter string to assert}

\item{response}{Name of response column}
}
\value{
String for eval and parsing
}
\description{
This function parses a \code{obs_filter} argument passed through a \code{predict_...}
function into a string ready to be evaluated in \code{merge_prediction()}. The output
string will be parsed within a \code{case_when}. The \code{obs_filter} passed sets a
threshold for fitting models and generating predictions for data based on
number of observed data points. So, \code{'>=5'} would be that we only want to fit a model,
generate predictions and use them them to replace observations
if there are 5 or more observations. Thus, in \code{case_when()}, we pass the first
logical argument to be anywhere there is NOT \code{'>=5'} observations.
}
