% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_filter.R
\name{parse_replace_filter}
\alias{parse_replace_filter}
\title{Parse replace filter intro string to be evaluated}
\usage{
parse_replace_filter(replace_filter, response)
}
\arguments{
\item{replace_filter}{Replace filter string to assert}

\item{response}{Name of response column}
}
\value{
String for eval and parsing
}
\description{
This function parses a \code{replace_filter} argument passed through a \code{predict_...}
function into a string ready to be evaluated in \code{merge_prediction()}. The output
string will be parsed within a \code{case_when}. The \code{replace_filter} passed sets a
threshold for replacing data based on number of observed data points. So,
\code{'>=5'} would be that we only want to use predictions and put them onto observations
if there are 5 or more observations. Thus, in \code{case_when()}, we pass the first
logical argument to be anywhere there is NOT \code{'>=5'} observations.
}
