% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_models.R
\name{predict_data}
\alias{predict_data}
\title{Generate prediction from model object}
\usage{
predict_data(df, model, pred_col, upper_col, lower_col)
}
\value{
A data frame.
}
\description{
\code{predict_data()} generates a prediction vector from a model object and full
data frame, putting this prediction back into the data frame.
}
