% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_models.R
\name{predict_forecast_data}
\alias{predict_forecast_data}
\title{Generate prediction from model object}
\usage{
predict_forecast_data(df, forecast_obj, pred_col, upper_col, lower_col)
}
\arguments{
\item{df}{Data frame of model data.}

\item{forecast_obj}{Object of class \code{forecast} that is output from the \verb{forecast::}
family of functions.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}
}
\value{
A data frame.
}
\description{
\code{predict_forecast_data()} generates a prediction vector from a forecast object
and full data frame, putting this prediction back into the data frame.
}
