% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_simple.R
\name{predict_simple}
\alias{predict_simple}
\title{Use linear interpolation and flat extrapolation to infill data}
\usage{
predict_simple(
  df,
  model = c("forward", "all", "flat_extrap", "linear_interp", "back_extrap",
    "both_extrap"),
  col = "value",
  ret = c("df", "all", "error"),
  test_col = NULL,
  test_period = NULL,
  test_period_flex = NULL,
  group_col = "iso3",
  obs_filter = NULL,
  sort_col = "year",
  sort_descending = FALSE,
  pred_col = "pred",
  type_col = NULL,
  types = c("imputed", "imputed", "projected"),
  source_col = NULL,
  source = NULL,
  scenario_detail_col = NULL,
  scenario_detail = NULL,
  replace_obs = c("missing", "none")
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{model}{Type of simple extrapolation or interpolation to perform:
\itemize{
\item{\code{forward}: }{Just \code{flat_extrap} and \code{linear_interp}. (default)}
\item{\code{all}: }{All of \code{flat_extrap}, \code{linear_interp}, and \code{back_extrap}}
\item{\code{flat_extrap}: }{Flat extrapolation from latest observed point.}
\item{\code{linear_interp}: }{Linear interpolation between observed data points.}
\item{\code{back_extrap}: }{Flat extrapolation from first observed data point backwards.}
\item{\code{both_extrap}: }{Both \code{flat_extrap} and \code{back_extrap}.}
}}

\item{col}{Name of column to extrapolate/interpolate.}

\item{ret}{Character vector specifying what values the function returns. Defaults
to returning a data frame, but can return a vector of model error, the
model itself or a list with all 3 as components.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored. See \code{\link[=model_error]{model_error()}}
for details on the methods and metrics returned.}

\item{test_period}{Length of period to test for RMChE. If \code{NULL}, beginning and end
points of each group in \code{group_col} are compared. Otherwise, \code{test_period} must
be set to an integer \code{n} and for each group, comparisons are made between
the end point and \code{n} periods prior.}

\item{test_period_flex}{Logical value indicating if \code{test_period} is less than
the full length of the series, should change error still be calculated for that
point. Defaults to \code{FALSE}.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.
If \code{NULL}, not used. Defaults to \code{"iso3"}.}

\item{obs_filter}{String value of the form "\verb{logical operator} \code{integer}"
that specifies the number of observations required to fit the model and
replace observations with predicted values. This is done in
conjunction with \code{group_col}. So, if \code{group_col = "iso3"} and
\code{obs_filter = ">= 5"}, then for this model, predictions will only be used
for \code{iso3} vales that have 5 or more observations. Possible logical operators
to use are \code{>}, \code{>=}, \code{<}, \code{<=}, \code{==}, and \code{!=}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{If `group_models = FALSE`, then `obs_filter` is only used to determine when
predicted values replace observed values but **is not** used to restrict values
from being used in model fitting. If `group_models = TRUE`, then a model
is only fit for a group if they meet the `obs_filter` requirements. This provides
speed benefits, particularly when running INLA time series using `predict_inla()`.
}\if{html}{\out{</div>}}}

\item{sort_col}{Column name(s) to use to dplyr::arrange() the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series. If NULL, not used. Defaults to "year". For \code{predict_simple()},
the first value in \code{sort_col} is passed to \code{\link[zoo:na.approx]{zoo::na.approx()}} as \code{xout} to
ensure linear interpolation is based on \code{sort_col} indexing rather than
default data frame indexing.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}

\item{pred_col}{Column name to store predicted value.}

\item{type_col}{Column name specifying data type.}

\item{types}{Vector of length 3 that provides the type to provide to data
produced in the model. These values are only used to fill in type values
where the dependent variable is missing. The first value is given to missing
observations that precede the first observation, the second to those after
the last observation, and the third for those following the final observation.}

\item{source_col}{Column name containing source information for the data frame.
If provided, the argument in \code{source} is used to fill in where predictions
have filled in missing data.}

\item{source}{Source to add to missing values.}

\item{scenario_detail_col}{Column name containing scenario_detail information
for the data frame. If provided, the argument in \code{scenario_detail} is used
to fill in where prediction shave filled in missing data.}

\item{scenario_detail}{Scenario details to add to missing values (usually the
name of the model being used to generate the projection, optionally with
relevant parameters).}

\item{replace_obs}{Character value specifying how, if at all, observations should
be replaced by infilled values. By default, replaces missing values in \code{col}
but if set to \code{"none"} then \code{col} is not changed.}
}
\value{
Depending on the value passed to \code{ret}, either a data frame with
predicted data, a vector of errors from \code{\link[=model_error]{model_error()}}, a fitted model, or a list with all 3.
}
\description{
\code{predict_simple()} does simple linear interpolation and/or flat extrapolation
on a column using \code{\link[zoo:na.approx]{zoo::na.approx()}}. Similar to other predict functions, it also
allows filling in of type and source if necessary. However, it does not provide
confidence bounds on the estimates, like other \code{predict_...} model-based
functions provide.
}
\details{
Depending on the value of \code{model} passed to the function, linear interpolation,
flat extrapolation, or both is used on the data.
}
