% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit_transform}
\alias{probit_transform}
\title{Probit transform bounded data in a data frame}
\usage{
probit_transform(df, cols, inverse = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{Names of columns to transform.}

\item{inverse}{Logical value indicating whether or not to inverse the transformation.}
}
\value{
A data frame.
}
\description{
\code{probit_transform()} transform specified columns in a data frame using
\code{\link[VGAM:probitlink]{VGAM::probitlink()}}. This is similar to a logit transformation but better
behaved at the extremes.
}
