% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit_vec}
\alias{probit_vec}
\title{Probit transformation for bounded data}
\usage{
probit_vec(x, inverse = FALSE)
}
\arguments{
\item{x}{Numeric vector to probit transform}

\item{inverse}{Logical value on whether to invert the transformation}
}
\value{
A numeric vector
}
\description{
Similar to logit transformation but better behaved at the extremes
}
