% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_groups.R
\name{remove_groups}
\alias{remove_groups}
\title{Remove groups from data frame if grouped}
\usage{
remove_groups(df)
}
\arguments{
\item{df}{Data frame.}
}
\description{
Groups from input data frames into \code{predict_...} functions are removed to prevent
errors when fitting data and bringing predictions to the data frame.
}
