% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_transform}
\alias{scale_transform}
\title{Scales data in a data frame}
\usage{
scale_transform(df, cols, scale = 100, inverse = FALSE)
}
\arguments{
\item{df}{A data frame.}

\item{cols}{Names of columns to transform.}

\item{inverse}{Logical value indicating whether or not to inverse the transformation.}
}
\value{
A data frame.
}
\description{
\code{scale_transform()} transform specified columns in a data frame using
a provided scale. If \code{inverse = FALSE}, it's simply \code{x * scale}, otherwise
it is \code{x / scale}.
}
