% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_vec}
\alias{scale_vec}
\title{Scale a vector}
\usage{
scale_vec(x, scale, divide)
}
\arguments{
\item{x}{Numeric vector.}

\item{scale}{Value to scale the vector(s) by. Defaults by 100}

\item{divide}{If \code{TRUE}, divide the vector by the scale. Otherwise, multiply.}
}
\description{
Scale a vector
}
