% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_simple.R
\name{simple_extrap}
\alias{simple_extrap}
\title{Helper function to do flat extrapolation}
\usage{
simple_extrap(x)
}
\arguments{
\item{x}{Vector to do flat extrapolation on}
}
\description{
Helper function to do flat extrapolation
}
