% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_average_trend.R
\name{temp_fill}
\alias{temp_fill}
\title{Fills vector backwards and forward, for use prior to applying average trend}
\usage{
temp_fill(x)
}
\arguments{
\item{x}{Vector to fill, typically response vector}
}
\description{
Fills vector backwards and forward, for use prior to applying average trend
}
