% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_models.R
\name{trim_series}
\alias{trim_series}
\title{Get latest data for forecasting}
\usage{
trim_series(x)
}
\arguments{
\item{x}{Data series to reduce for forecasting}
}
\value{
Series with contiguous observations followed by NA values to forecast.
}
\description{
Gets latest data for forecasting. It also gets the number of missing data
points to forecast.
}
