% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariates_df}
\alias{covariates_df}
\title{Default covariates for use in augury functions.}
\format{
A data frame with 5044 rows and 8 variables:
\describe{
\item{iso3}{Country ISO3 code.}
\item{year}{Year}
\item{region}{World Bank regional classification}
\item{sdi}{Socio-demographic Index, a composite calculated using publicly available on total fertility rates, mean years of schooling, and lag distributed income per capita.}
\item{e0}{Life expectancy at birth, total (years)}
}
}
\source{
\href{https://datahelpdesk.worldbank.org/knowledgebase/articles/906519-world-bank-country-and-lending-groups}{Region, World Bank regional groups, 2020.}

\href{https://gpw13.github.io/#methods-for-infilling-and-projections}{Socio-demographic Index, World Health Organization calculation, 2020.}

\href{https://population.un.org/wpp/Download/Standard/Mortality/}{Life expectancy, World Population Prospects, 2019.}
}
\usage{
covariates_df
}
\description{
A dataset containing default covariates for all WHO member states, for use
within augury functions. These are used if no defaults are provided
to the modeling functions.
}
\keyword{datasets}
