% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_models.R
\name{fit_inla_model}
\alias{fit_inla_model}
\title{Fit INLA model to data}
\usage{
fit_inla_model(
  df,
  formula,
  control.predictor,
  ...,
  formula_vars,
  test_col,
  group_col,
  group_models,
  pred_col,
  upper_col,
  lower_col,
  filter_na,
  ret
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{formula}{A formula that will be supplied to the model, such as \code{y~x}.}

\item{control.predictor}{Used to set \code{compute = TRUE} to ensure that the posterior
marginals of the fitted values are obtained and the mean and standard deviation
of the fitted values returned for use in the infilling and predictions. Additional
arguments can be passed in the \code{control.predictor} list, but must always include
\code{compute = TRUE}. See \code{\link[INLA:control.predictor]{INLA::control.predictor()}} for details.}

\item{...}{Additional arguments passed to \code{\link[INLA:inla]{INLA::inla()}}.}

\item{formula_vars}{Variables included in the model formula, generated by
\code{all.vars(formula)}.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.}

\item{group_models}{Logical, if \code{TRUE}, fits and predicts models individually onto
each \code{group_col}. If \code{FALSE}, a general model is fit across the entire data
frame.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}

\item{filter_na}{Character value specifying how, if at all, to filter \code{NA}
values from the dataset prior to applying the model. By default, only
observations with missing predictors are removed, although it can also remove
rows only if they have missing dependent or independent variables, or no
filtering at all. Model prediction and fitting are done in one pass with
\code{\link[INLA:inla]{INLA::inla()}}, so there will be no predictions if observations with
missing dependent variables are removed.}

\item{ret}{Character vector specifying what values the function returns. Defaults
to returning a data frame, but can return a vector of model error, the
model itself or a list with all 3 as components.}
}
\value{
List of \code{mdl} (fitted model) and \code{df} (data frame with fitted values
and confidence bounds generated from the model).
}
\description{
Used within \code{predict_inla()}, this function fits the model to the data
frame, working whether the model is being fit across the entire data frame or
being fit to each group individually. Data is filtered prior to fitting,
model(s) are fit, and then fitted values are generated on the original.
}
\details{
If fitting models individually to each group, \code{mdl} will never be returned, as
as these are instead a large group of models. Otherwise, a list of \code{mdl} and \code{df}
is returned and used within \code{predict_inla()}.
}
