% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_data.R
\name{get_model_data}
\alias{get_model_data}
\title{Minimizes dataset to data needed for modelling}
\usage{
get_model_data(
  df,
  formula_vars,
  test_col,
  group_col = NULL,
  filter_na,
  reduce_columns = TRUE
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{formula_vars}{Character vector of variables used in the model. Can be
extracted from a formula using \code{all.vars(fmla)}.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored.}

\item{group_col}{Column name to split data frame prior to model application.}

\item{filter_na}{Character value specifying how, if at all, to filter \code{NA}
values from the dataset prior to applying the model.  By default, all
observations with missing values are removed, although it can also remove
rows only if they have missing dependent or independent variables, or no
filtering at all.}

\item{reduce_columns}{Logical on whether or not to reduce the number of columns
in the data to just those necessary for modelling.}
}
\value{
A data frame.
}
\description{
\code{get_model_data()} ensures that only variables necessary for the model
are included in the dataset and missing data and test sets are removed, if
\code{test_col} is not \code{NULL}. If \code{filter_na} is \code{"all"} (the default), then any
observations with \code{NA} values are removed using \code{na.omit()}. If \code{filter_na} is
\code{"response"} or \code{"predictors"} then only rows with missing dependent or independent
variables are removed, respectively. If \code{"none"}, then no filtering is done at all.
}
