% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_inla.R
\name{map_model_behavior}
\alias{map_model_behavior}
\title{Catch instability of INLA}
\usage{
map_model_behavior(df, formula, control.predictor, ...)
}
\arguments{
\item{df}{Data frame of model data.}

\item{formula}{A formula that will be supplied to the model, such as \code{y~x}.}

\item{control.predictor}{Used to set \code{compute = TRUE} to ensure that the posterior
marginals of the fitted values are obtained and the mean and standard deviation
of the fitted values returned for use in the infilling and predictions. Additional
arguments can be passed in the \code{control.predictor} list, but must always include
\code{compute = TRUE}. See \code{\link[INLA:control.predictor]{INLA::control.predictor()}} for details.}

\item{...}{Additional arguments passed to \code{\link[INLA:inla]{INLA::inla()}}.}
}
\value{
A data frame.
}
\description{
\code{map_model_behavior()} runs \code{\link[INLA:inla]{INLA::inla()}} until its successful. This is
important on  linux machines where INLA can produce unstable results.
}
