% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_average_trend.R
\name{merge_average_df}
\alias{merge_average_df}
\title{Merge average df with predictions with original data frame}
\usage{
merge_average_df(
  avg_df,
  df,
  response,
  average_cols,
  group_col,
  sort_col,
  pred_col,
  upper_col,
  lower_col
)
}
\arguments{
\item{avg_df}{Data frame with average trends.}

\item{df}{Data frame of model data.}

\item{response}{Column name of response variable.}

\item{average_cols}{Column name(s) of column(s) for use in grouping data for averaging,
such as regions. If missing, uses global average of the data for infilling.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}
}
\value{
Original data frame with new trend joined up.
}
\description{
Merge average df with predictions with original data frame
}
