% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_merge.R
\name{merge_prediction}
\alias{merge_prediction}
\title{Merge predicted data into data frame}
\usage{
merge_prediction(
  df,
  response,
  pred_col,
  upper_col,
  lower_col,
  type_col,
  types,
  type_group,
  type_sort,
  source_col,
  source,
  replace_obs,
  error_correct,
  error_correct_cols
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}

\item{type_col}{Column name specifying data type.}

\item{types}{Vector of length 3 that provides the type to provide to data
produced in the model. These values are only used to fill in type values
where the dependent variable is missing. The first value is given to missing
observations that precede the first observation, the second to those after
the last observation, and the third for those following the final observation.}

\item{type_group}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type.}

\item{type_sort}{Column name to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type.}

\item{replace_obs}{Character value specifying how, if at all, observations should
be replaced by fitted values. Defaults to replacing only missing values,
but can be used to replace all values or none.}

\item{error_correct}{Logical value indicating whether or not whether mean error
should be used to adjust predicted values. If \code{TRUE}, the mean error between
observed and predicted data points will be used to adjust predictions. If
\code{error_correct_cols} is not \code{NULL}, mean error will be used within those
groups instead of overall mean error.}

\item{error_correct_cols}{Column names of data frame to group by when applying
error correction to the predicted values.}
}
\value{
A data frame.
}
\description{
Merges predicted data into data frame. By default, does not replace observed
values with modeled data, although upper and lower bounds for the model always
included.
}
