% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_prediction.R
\name{merge_prediction}
\alias{merge_prediction}
\title{Merge predicted data into data frame}
\usage{
merge_prediction(
  df,
  response,
  group_col,
  sort_col,
  sort_descending,
  pred_col,
  type_col,
  types,
  source_col,
  source,
  replace_obs,
  replace_filter
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{response}{Column name of response variable.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.
If \code{NULL}, not used. Defaults to \code{"iso3"}.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series. If \code{NULL}, not used. Defaults to \code{"year"}.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}

\item{pred_col}{Column name to store predicted value.}

\item{type_col}{Column name specifying data type.}

\item{types}{Vector of length 3 that provides the type to provide to data
produced in the model. These values are only used to fill in type values
where the dependent variable is missing. The first value is given to missing
observations that precede the first observation, the second to those after
the last observation, and the third for those following the final observation.}

\item{source_col}{Column name containing source information for the data frame.
If provided, the argument in \code{source} is used to fill in where predictions
have filled in missing data.}

\item{source}{Source to add to missing values.}

\item{replace_obs}{Character value specifying how, if at all, observations should
be replaced by fitted values. Defaults to replacing only missing values,
but can be used to replace all values or none.}

\item{replace_filter}{String value of the form "\verb{logical operator} \code{integer}"
that specifies when replacing observations by predicted values, this is
done where there is a specific number of observations. This is done in
conjunction with \code{group_col}. So, if \code{group_col = "iso3"} and
\code{replace_filter = ">= 5"}, then for this model, predictions will only be used
for \code{iso3} vales that have 5 or more observations. Possible logical operators
to use are \code{>}, \code{>=}, \code{<}, \code{<=}, \code{==}, and \code{!=}.}
}
\value{
A data frame.
}
\description{
Merges predicted data into data frame. By default, does not replace observed
values with modeled data.
}
