% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_average.R
\name{predict_average_fn}
\alias{predict_average_fn}
\title{Impute data using simple averages}
\usage{
predict_average_fn(
  df,
  col,
  average_cols = NULL,
  weight_col = NULL,
  flat_extrap = TRUE,
  test_col = NULL,
  group_col = NULL,
  pred_col = "pred",
  sort_col = NULL,
  sort_descending = FALSE,
  error_correct = FALSE,
  error_correct_cols = NULL
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{col}{Name of column to extrapolate/interpolate.}

\item{average_cols}{Column name(s) of column(s) for use in grouping data for averaging,
such as regions. If missing, uses global average of the data for infilling.}

\item{weight_col}{Column name of column of weights to be used in averaging, such
as country population.}

\item{flat_extrap}{Logical value determining whether or not to flat extrapolate
using the latest average for missing rows with no data available.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.}

\item{pred_col}{Column name to store predicted value.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}

\item{error_correct}{Logical value indicating whether or not whether mean error
should be used to adjust predicted values. If \code{TRUE}, the mean error between
observed and predicted data points will be used to adjust predictions. If
\code{error_correct_cols} is not \code{NULL}, mean error will be used within those
groups instead of overall mean error.}

\item{error_correct_cols}{Column names of data frame to group by when applying
error correction to the predicted values.}
}
\value{
A data frame.
}
\description{
\code{predict_average_fn()} does simple imputation and flat extrapolation
using averages grouped by \code{average_cols}.
}
