% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_forecast.R
\name{predict_forecast_data}
\alias{predict_forecast_data}
\title{Generate prediction from model object}
\usage{
predict_forecast_data(
  df,
  forecast_obj,
  sort_col,
  sort_descending,
  pred_col,
  upper_col,
  lower_col
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{forecast_obj}{Object of class \code{forecast} that is output from the \verb{forecast::}
family of functions.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series. If \code{NULL}, not used. Defaults to \code{"year"}.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}
}
\value{
A data frame.
}
\description{
\code{predict_forecast_data()} generates a prediction vector from a forecast object
and full data frame, putting this prediction back into the data frame.
}
