% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_models.R
\name{predict_inla}
\alias{predict_inla}
\title{Use Bayesian analysis of additive models to infill and project data}
\usage{
predict_inla(
  df,
  formula,
  ...,
  ret = c("df", "all", "error", "model"),
  test_col = NULL,
  pred_col = "pred",
  upper_col = "upper",
  lower_col = "lower",
  filter_na = c("predictors", "response", "all", "none"),
  type_col = NULL,
  types = c("imputed", "imputed", "projected"),
  type_group = "iso3",
  type_sort = "year",
  source_col = NULL,
  source = NULL,
  replace_obs = c("missing", "all", "none"),
  error_correct = FALSE,
  error_correct_cols = NULL
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{formula}{A formula that will be supplied to the model, such as \code{y~x}.}

\item{...}{Additional arguments passed to \code{\link[INLA:inla]{INLA::inla()}}.}

\item{ret}{Character vector specifying what values the function returns. Defaults
to returning a data frame, but can return a vector of model error, the
model itself or a list with all 3 as components.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}

\item{filter_na}{Character value specifying how, if at all, to filter \code{NA}
values from the dataset prior to applying the model. By default, only
observations with missing predictors are removed, although it can also remove
rows only if they have missing dependent or independent variables, or no
filtering at all. Model prediction and fitting are done in one pass with
\code{\link[INLA:inla]{INLA::inla()}}, so there will be no predictions if observations with
missing dependent variables are removed.}

\item{type_col}{Column name specifying data type.}

\item{types}{Vector of length 3 that provides the type to provide to data
produced in the model. These values are only used to fill in type values
where the dependent variable is missing. The first value is given to missing
observations that precede the first observation, the second to those after
the last observation, and the third for those following the final observation.}

\item{type_group}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type.}

\item{type_sort}{Column name to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type.}

\item{replace_obs}{Character value specifying how, if at all, observations should
be replaced by fitted values. Defaults to replacing only missing values,
but can be used to replace all values or none.}

\item{error_correct}{Logical value indicating whether or not whether mean error
should be used to adjust predicted values. If \code{TRUE}, the mean error between
observed and predicted data points will be used to adjust predictions. If
\code{error_correct_cols} is not \code{NULL}, mean error will be used within those
groups instead of overall mean error.}

\item{error_correct_cols}{Column names of data frame to group by when applying
error correction to the predicted values.}
}
\value{
Depending on the value passed to \code{ret}, either a data frame with
predicted data, a vector of errors, a fitted model, or a list with all 3.
}
\description{
\code{predict_inla()} uses a Integrated Nested Laplace approximation to fit a model
and use that model to infill and project the dependent variable. The function
uses \code{\link[INLA:inla]{INLA::inla()}} to perform the model fitting and prediction, and full
details and explanation of arguments that it can accept is available on that page.
}
\details{
The function also allows for inputting of data type and source information
directly into the data frame if the \code{type_col} and \code{source_col} are specified
respectively.
}
