% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_average_model.R
\name{predict_inla_avg_trend}
\alias{predict_inla_avg_trend}
\title{Use \code{predict_inla} on groups to generate average trend and apply to original data}
\usage{
predict_inla_avg_trend(
  df,
  formula,
  average_cols = NULL,
  weight_col = NULL,
  group_models = FALSE,
  control.predictor = list(compute = TRUE),
  ...,
  ret = c("df", "all", "error", "model"),
  scale = NULL,
  probit = FALSE,
  test_col = NULL,
  test_period = NULL,
  test_period_flex = NULL,
  group_col = NULL,
  sort_col = NULL,
  sort_descending = FALSE,
  pred_col = "pred",
  upper_col = "upper",
  lower_col = "lower",
  filter_na = c("predictors", "response", "all", "none"),
  type_col = NULL,
  types = c("imputed", "imputed", "projected"),
  source_col = NULL,
  source = NULL,
  replace_obs = c("missing", "all", "none"),
  error_correct = FALSE,
  error_correct_cols = NULL,
  shift_trend = FALSE
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{formula}{A formula that will be supplied to the model, such as \code{y~x}.
Variables defined in the formula will be used in the averaging. If the
variable is defined as part of \code{average_cols}, then it will be used within
\code{\link[dplyr:group_by]{dplyr::group_by()}} prior to averaging. If it is not a part of \code{average_cols},
then it must be a numeric column whose average will be taken.}

\item{average_cols}{Column name(s) of column(s) for use in grouping data for averaging,
such as regions. If missing, uses global average of the data for infilling.}

\item{weight_col}{Column name of column of weights to be used in averaging, such
as country population.}

\item{group_models}{Logical, whether or not to run separate models for each group
defined by \code{average_cols}. If the \code{sort_col} is part of \code{average_cols}, then
it is not used to group models.}

\item{control.predictor}{Used to set \code{compute = TRUE} to ensure that the posterior
marginals of the fitted values are obtained and the mean and standard deviation
of the fitted values returned for use in the infilling and predictions. Additional
arguments can be passed in the \code{control.predictor} list, but must always include
\code{compute = TRUE}. See \code{\link[INLA:control.predictor]{INLA::control.predictor()}} for details.}

\item{...}{Additional arguments passed to \code{\link[INLA:inla]{INLA::inla()}}.}

\item{ret}{Character vector specifying what values the function returns. Defaults
to returning a data frame, but can return a vector of model error, the
model itself or a list with all 3 as components.}

\item{scale}{Either \code{NULL} or a numeric value. If a numeric value is provided,
the response variable is scaled by the value passed to scale prior to model
fitting and prior to any probit transformation, so can be used to put the
response onto a 0 to 1 scale. Scaling is done by dividing the response by
the scale and using the \code{\link[=scale_transform]{scale_transform()}} function. The response, as well
as the fitted values and confidence bounds are unscaled prior to error
calculation and returning to the user.}

\item{probit}{Logical value on whether or not to probit transform the response
prior to model fitting. Probit transformation is performed after any scaling
determined by \code{scale} but prior to model fitting. The response, as well as
the fitted values and confidence bounds are untransformed prior to error
calculation and returning to the user.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored. See \code{\link[=model_error]{model_error()}}
for details on the methods and metrics returned.}

\item{test_period}{Length of period to test for RMChE. If \code{NULL}, beginning and end
points of each group in \code{group_col} are compared. Otherwise, \code{test_period} must
be set to an integer \code{n} and for each group, comparisons are made between
the end point and \code{n} periods prior.}

\item{test_period_flex}{Logical value indicating if \code{test_period} is less than
the full length of the series, should change error still be calculated for that
point. Defaults to \code{FALSE}.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}

\item{filter_na}{Character value specifying how, if at all, to filter \code{NA}
values from the dataset prior to applying the model. By default, only
observations with missing predictors are removed, although it can also remove
rows only if they have missing dependent or independent variables, or no
filtering at all. Model prediction and fitting are done in one pass with
\code{\link[INLA:inla]{INLA::inla()}}, so there will be no predictions if observations with
missing dependent variables are removed.}

\item{type_col}{Column name specifying data type.}

\item{types}{Vector of length 3 that provides the type to provide to data
produced in the model. These values are only used to fill in type values
where the dependent variable is missing. The first value is given to missing
observations that precede the first observation, the second to those after
the last observation, and the third for those following the final observation.}

\item{source_col}{Column name containing source information for the data frame.
If provided, the argument in \code{source} is used to fill in where predictions
have filled in missing data.}

\item{source}{Source to add to missing values.}

\item{replace_obs}{Character value specifying how, if at all, observations should
be replaced by fitted values. Defaults to replacing only missing values,
but can be used to replace all values or none.}

\item{error_correct}{Logical value indicating whether or not whether mean error
should be used to adjust predicted values. If \code{TRUE}, the mean error between
observed and predicted data points will be used to adjust predictions. If
\code{error_correct_cols} is not \code{NULL}, mean error will be used within those
groups instead of overall mean error.}

\item{error_correct_cols}{Column names of data frame to group by when applying
error correction to the predicted values.}

\item{shift_trend}{Logical value specifying whether or not to shift predictions
so that the trend matches up to the last observation. If \code{error_correct} and
\code{shift_trend} are both \code{TRUE}, \code{shift_trend} takes precedence.}
}
\value{
Depending on the value passed to \code{ret}, either a data frame with
predicted data, a vector of errors from \code{\link[=model_error]{model_error()}}, a fitted model, or a list with all 3.
}
\description{
\code{predict_inla_avg_trend()} uses a Integrated Nested Laplace approximation to fit a model
to groups within the data, and then bring that fitted prediction back to the
original data. The function uses \code{\link[INLA:inla]{INLA::inla()}} to perform the model fitting
and prediction, and full details and explanation of arguments that it can accept is available on that page.
The function also allows for inputting of data type and source information
directly into the data frame if the \code{type_col} and \code{source_col} are specified
respectively.
}
\details{
\code{predict_..._avg_trend()} functions need to be used carefully. Ensure that \code{average_cols}
and variables in the \code{formula} match, and any \code{formula} variables not in \code{average_cols}
are numeric that can be averaged. Even though the modeling won't use the \code{group_col},
it should be provided if necessary to be used in error metric calculations, and provision
of \code{types} into \code{type_col}. Similarly, the \code{sort_col} is necessary for \code{types}, but
\strong{also needs to be in \code{average_cols}} if \code{error_correct}, \code{group_models}, or \code{shift_trend} is
going to be used.
}
