% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_models.R
\name{predict_inla_data}
\alias{predict_inla_data}
\title{Generate prediction from an INLA output object}
\usage{
predict_inla_data(df, model, pred_col, upper_col, lower_col)
}
\value{
A data frame.
}
\description{
\code{predict_inla_data()} generates a prediction vector from an \code{\link[INLA:inla]{INLA::inla()}}
output object, putting this prediction back into the data frame.
}
