% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_wrappers.R
\name{predict_inla_me}
\alias{predict_inla_me}
\title{Use INLA for mixed effects modeling for prediction}
\usage{
predict_inla_me(
  df,
  formula = value ~ sdi_scaled + e0_scaled + INLA::f(region, model = "iid") +
    INLA::f(year_n, model = "rw2"),
  family = "gaussian",
  control.predictor = list(compute = TRUE),
  control.inla = list(strategy = "laplace"),
  ...,
  ret = c("df", "all", "error", "model"),
  test_col = NULL,
  pred_col = "pred",
  upper_col = "upper",
  lower_col = "lower",
  filter_na = c("predictors", "response", "all", "none"),
  type_col = NULL,
  types = c("imputed", "imputed", "projected"),
  type_group = "iso3",
  type_sort = "year",
  source_col = NULL,
  source = NULL,
  replace_obs = c("missing", "all", "none"),
  error_correct = TRUE,
  error_correct_cols = "iso3"
)
}
\description{
Use INLA for mixed effects modeling for prediction
}
