% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_model_wrappers.R
\name{predict_inla_me}
\alias{predict_inla_me}
\title{Use INLA for mixed effects modeling for prediction}
\usage{
predict_inla_me(
  df,
  model = "rw2",
 
    formula = stats::as.formula(sprintf("value ~ sdi_scaled + e0_scaled + f(region, model = 'iid') + f(year_n, model = '\%s')",
    model)),
  control.predictor = list(compute = TRUE),
  family = "gaussian",
  control.inla = list(strategy = "laplace"),
  ...,
  ret = c("df", "all", "error", "model"),
  scale = NULL,
  probit = FALSE,
  test_col = NULL,
  group_col = "iso3",
  group_models = FALSE,
  sort_col = "year",
  sort_descending = FALSE,
  pred_col = "pred",
  upper_col = "upper",
  lower_col = "lower",
  filter_na = c("predictors", "response", "all", "none"),
  type_col = NULL,
  types = c("imputed", "imputed", "projected"),
  source_col = NULL,
  source = NULL,
  replace_obs = c("missing", "all", "none"),
  error_correct = TRUE,
  error_correct_cols = "iso3"
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{model}{Argument passed to \code{\link[INLA:f]{INLA::f()}} to define the time series model
used in the default formula. Defaults to "rw2".}

\item{formula}{A formula that will be supplied to the model, such as \code{y~x}.}

\item{control.predictor}{Used to set \code{compute = TRUE} to ensure that the posterior
marginals of the fitted values are obtained and the mean and standard deviation
of the fitted values returned for use in the infilling and predictions. Additional
arguments can be passed in the \code{control.predictor} list, but must always include
\code{compute = TRUE}. See \code{\link[INLA:control.predictor]{INLA::control.predictor()}} for details.}

\item{family}{A string indicating the likelihood family. The default is
\code{gaussian} with identity link. See \code{names(INLA::inla.model()$likelihood)}
for a list of possible alternatives and use \code{\link[INLA:doc]{INLA::inla.doc()}} for
detailed docs for individual families.}

\item{control.inla}{See \code{\link[INLA:control.inla]{INLA::control.inla()}} for details.}

\item{...}{Additional arguments passed to \code{\link[INLA:inla]{INLA::inla()}}.}

\item{ret}{Character vector specifying what values the function returns. Defaults
to returning a data frame, but can return a vector of model error, the
model itself or a list with all 3 as components.}

\item{scale}{Either \code{NULL} or a numeric value. If a numeric value is provided,
the response variable is scaled by the value passed to scale prior to model
fitting and prior to any probit transformation, so can be used to put the
response onto a 0 to 1 scale. Scaling is done by dividing the response by
the scale and using the \code{\link[=scale_transform]{scale_transform()}} function. The response, as well
as the fitted values and confidence bounds are unscaled prior to error
calculation and returning to the user.}

\item{probit}{Logical value on whether or not to probit transform the response
prior to model fitting. Probit transformation is performed after any scaling
determined by \code{scale} but prior to model fitting. The response, as well as
the fitted values and confidence bounds are untransformed prior to error
calculation and returning to the user.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored. See \code{\link[=model_error]{model_error()}}
for details on the methods and metrics returned.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.}

\item{group_models}{Logical, if \code{TRUE}, fits and predicts models individually onto
each \code{group_col}. If \code{FALSE}, a general model is fit across the entire data
frame.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}

\item{pred_col}{Column name to store predicted value.}

\item{upper_col}{Column name to store upper bound of confidence interval.}

\item{lower_col}{Column name to store lower bound of confidence interval.}

\item{filter_na}{Character value specifying how, if at all, to filter \code{NA}
values from the dataset prior to applying the model. By default, only
observations with missing predictors are removed, although it can also remove
rows only if they have missing dependent or independent variables, or no
filtering at all. Model prediction and fitting are done in one pass with
\code{\link[INLA:inla]{INLA::inla()}}, so there will be no predictions if observations with
missing dependent variables are removed.}

\item{type_col}{Column name specifying data type.}

\item{types}{Vector of length 3 that provides the type to provide to data
produced in the model. These values are only used to fill in type values
where the dependent variable is missing. The first value is given to missing
observations that precede the first observation, the second to those after
the last observation, and the third for those following the final observation.}

\item{source_col}{Column name containing source information for the data frame.
If provided, the argument in \code{source} is used to fill in where predictions
have filled in missing data.}

\item{source}{Source to add to missing values.}

\item{replace_obs}{Character value specifying how, if at all, observations should
be replaced by fitted values. Defaults to replacing only missing values,
but can be used to replace all values or none.}

\item{error_correct}{Logical value indicating whether or not whether mean error
should be used to adjust predicted values. If \code{TRUE}, the mean error between
observed and predicted data points will be used to adjust predictions. If
\code{error_correct_cols} is not \code{NULL}, mean error will be used within those
groups instead of overall mean error.}

\item{error_correct_cols}{Column names of data frame to group by when applying
error correction to the predicted values.}
}
\value{
Depending on the value passed to \code{ret}, either a data frame with
predicted data, a vector of errors from \code{\link[=model_error]{model_error()}}, a fitted model, or a list with all 3.
}
\description{
\code{predict_inla_me()} uses a Integrated Nested Laplace approximation to
fit a model and use that model to infill and project the dependent variable. See
\code{\link[=predict_inla]{predict_inla()}} for more details, but the function defaults to fitting
a mixed-effects model with the Socio-Demographic Index and life-expectancy at
birth (both scaled) used as covariates and second-order random walk time
process incorporated.
}
