% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_simple.R
\name{predict_simple_fn}
\alias{predict_simple_fn}
\title{Linearly interpolate data}
\usage{
predict_simple_fn(
  df,
  model,
  col,
  test_col = NULL,
  group_col = NULL,
  obs_filter = NULL,
  pred_col = "pred",
  sort_col = NULL,
  sort_descending = FALSE
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{model}{Type of simple extrapolation or interpolation to perform:
\itemize{
\item{\code{forward}: }{Just \code{flat_extrap} and \code{linear_interp}. (default)}
\item{\code{all}: }{All of \code{flat_extrap}, \code{linear_interp}, and \code{back_extrap}}
\item{\code{flat_extrap}: }{Flat extrapolation from latest observed point.}
\item{\code{linear_interp}: }{Linear interpolation between observed data points.}
\item{\code{back_extrap}: }{Flat extrapolation from first observed data point backwards.}
\item{\code{both_extrap}: }{Both \code{flat_extrap} and \code{back_extrap}.}
}}

\item{col}{Name of column to extrapolate/interpolate.}

\item{test_col}{Name of logical column specifying which response values to remove
for testing the model's predictive accuracy. If \code{NULL}, ignored. See \code{\link[=model_error]{model_error()}}
for details on the methods and metrics returned.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.
If \code{NULL}, not used. Defaults to \code{"iso3"}.}

\item{obs_filter}{String value of the form "\verb{logical operator} \code{integer}"
that specifies the number of observations required to fit the model and
replace observations with predicted values. This is done in
conjunction with \code{group_col}. So, if \code{group_col = "iso3"} and
\code{obs_filter = ">= 5"}, then for this model, predictions will only be used
for \code{iso3} vales that have 5 or more observations. Possible logical operators
to use are \code{>}, \code{>=}, \code{<}, \code{<=}, \code{==}, and \code{!=}.}

\item{pred_col}{Column name to store predicted value.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series. If \code{NULL}, not used. Defaults to \code{"year"}.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}
}
\value{
A data frame.
}
\description{
\code{predict_simple_fn()} does simple linear interpolation and flat extrapolation
on specified columnn in a data frame using \code{\link[zoo:na.approx]{zoo::na.approx()}}.
}
