% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_simple.R
\name{predict_simple_fn}
\alias{predict_simple_fn}
\title{Linearly interpolate data}
\usage{
predict_simple_fn(
  df,
  model,
  col,
  group_col = NULL,
  pred_col = "pred",
  sort_col = NULL,
  sort_descending = FALSE
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{model}{An R function that outputs a model object with a \code{predict.model} generic,
where \code{\link[stats:family]{stats::family()}} contains an inverse link function \code{linkinv} and
\code{predict.model()} accepts the \code{se.fit = TRUE} argument and returns confidence
intervals. This includes \link[stats:lm]{stats::lm}, \link[stats:glm]{stats::glm}, and \link[lme4:lmer]{lme4::lmer}.}

\item{col}{Name of column to extrapolate/interpolate.}

\item{group_col}{Column name(s) of group(s) to use in \code{\link[dplyr:group_by]{dplyr::group_by()}} when
supplying type, calculating mean absolute scaled error on data involving
time series, and if \code{group_models}, then fitting and predicting models too.}

\item{pred_col}{Column name to store predicted value.}

\item{sort_col}{Column name(s) to use to \code{\link[dplyr:arrange]{dplyr::arrange()}} the data prior to
supplying type and calculating mean absolute scaled error on data involving
time series.}

\item{sort_descending}{Logical value on whether the sorted values from \code{sort_col}
should be sorted in descending order. Defaults to \code{FALSE}.}
}
\value{
A data frame.
}
\description{
\code{predict_simple_fn()} does simple linear interpolation and flat extrapolation
on specified columnn in a data frame using \code{\link[zoo:na.approx]{zoo::na.approx()}}.
}
