% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_simple.R
\name{predict_simple_fn}
\alias{predict_simple_fn}
\title{Linearly interpolate data}
\usage{
predict_simple_fn(
  df,
  model,
  col,
  group_col = NULL,
  pred_col = "pred",
  sort_col = NULL,
  sort_descending = FALSE
)
}
\arguments{
\item{df}{Data frame of model data.}

\item{model}{An R function that outputs a model object with a \code{predict.model} generic,
where \code{\link[stats:family]{stats::family()}} contains an inverse link function \code{linkinv} and
\code{predict.model()} accepts the \code{se.fit = TRUE} argument and returns confidence
intervals. This includes \link[stats:lm]{stats::lm}, \link[stats:glm]{stats::glm}, and \link[lme4:lmer]{lme4::lmer}.}

\item{col}{Name of column to extrapolate/interpolate.}

\item{group_col}{Column name to split data frame prior to model application.}

\item{pred_col}{Column name to store predicted value.}

\item{sort_col}{Name of column to arrange data frame by prior to interpolation. Defaults to \code{"year"}.}

\item{sort_descending}{Logical value on whether the sorted values should be descending. Defaults to \code{FALSE}.}
}
\value{
A data frame.
}
\description{
\code{linear_interpolation()} does simple linear interpolation on specified columns
in a data frame using \code{\link[zoo:na.approx]{zoo::na.approx()}}.
}
