% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_simple.R
\name{simple_extrap}
\alias{simple_extrap}
\title{Helper function to do flat extrapolation}
\usage{
simple_extrap(x, model)
}
\arguments{
\item{x}{Vector to do flat extrapolation on}

\item{model}{Type of extrapolation to do (backward or forward)}
}
\description{
Helper function to do flat extrapolation
}
